Move a lot of the I/O routines to hashtable-common, so they can be reused by the sparsehashtable code.
To make this easier, changed some code that had taken a
size_type, to take a size_t as well.  Should be equivalent.

I now document that the INPUT/OUTPUT parameters must either be
a FILE* or have a Read/Write method on them.  If you want to
serialize to a different type, such as a string, you will need
to provide a wrapper that implements Read/Write.  I did so with
the string class in hashtable_test, as a demonstration.

I realized that everything in hashtable-common was in the
global namespace, so I fixed
that.  I put this stuff in a new sparsehash_internal
namespace, so it's less likely to yield namespace collisions.

R=igorginzburg
DELTA=297  (192 added, 71 deleted, 34 changed)


Revision created by MOE tool push_codebase.
MOE_MIGRATION=3561


git-svn-id: http://sparsehash.googlecode.com/svn/trunk@88 21bedea4-f223-4c8b-73d6-85019ffb75a9