Add support for serializing to and from a C++ stream, based on an idea from ddigennaro.  We overload read_data/write_data to support stream arguments, but it's tricky:
For technical reasons, we implement read_data/write_data in two
stages.  The actual work is done in *_data_internal, which takes
the stream argument twice: once as a template type, and once with
normal type information.  (We only use the second version.)  We do
this because of how C++ picks what function overload to use.  If we
implemented this the naive way:
bool read_data(istream* is, const void* data, size_t length);
template<typename T> read_data(T* fp,  const void* data, size_t length);
C++ would prefer the second version for every stream type except
istream.  However, we want C++ to prefer the first version for
streams that are *subclasses* of istream, such as strstream.  This
is not possible given the way template types are resolved.  So we
split the stream argument in two, one of which is templated and one
of which is not.  The specialized functions (like the istream
version above) ignore the template arg and use the second, 'type'
arg, getting subclass matching as normal.  The 'catch-all'
functions (the second version above) use the template arg to deduce
the type, and use a second, void* arg to achieve the desired
'catch-all' semantics.

The other option would have been to try to pass iostreams by
value rather than by pointer (so the 'catch-all' overload
wouldn't match for streams).  This runs into trouble of its
own, because you can't pass iostreams by value -- they're not
copyable, sensibly -- but by default the templated functions
will try to pass things around by value.  So we need to put in
explicit template args indicating things are by reference, and
the code gets really ugly, with code duplication in a lot of
places and code like 'hasher.operator()<T&>(foo);'.  This
solution is much cleaner, even if it does mean you have to
pass in iostreams by pointer, which is contra normal C++ use.

I added an example to hashtable_test to test this new
functionality.

R=igorginzburg,austern
DELTA=260  (162 added, 76 deleted, 22 changed)


Revision created by MOE tool push_codebase.
MOE_MIGRATION=3601


git-svn-id: http://sparsehash.googlecode.com/svn/trunk@89 21bedea4-f223-4c8b-73d6-85019ffb75a9